feat(auth): refactor login and error pages to use new components and improve localization support
